requirements.txt
abelectronics @ git+https://github.com/abelectronicsuk/ABElectronics_Python_Libraries.git@b9a85e418ab6a6701344c7e7d9a243d698cbebaa
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
firebase-admin==6.5.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
numpy==2.0.2
pandas==2.2.3
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
RPi.bme280==0.2.4
rsa==4.9
six==1.16.0
smbus2==0.4.3
spidev==3.6
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
